chore: move Arr into Support
